refactor: update lib modules for new architecture
